Add image metadata buffer

+++ src/example/images/images.cpp
@@ -276,7 +276,7 @@ int main( int argc, const char *argv[] ) {
     auto c = cdist( rng );
     auto d = ddist( rng );
     window.add_rectangle(
-      vk::Rect2D{ vk::Offset2D{ x0, y0 }, vk::Extent2D{ x1 - x0, y1 - y0 } },
+      vk::Rect2D{ vk::Offset2D{ x0, y0 }, vk::Extent2D{ static_cast< unsigned int >( x1 - x0 ), static_cast< unsigned int >( y1 - y0 ) } },
       gct::srgb_oetf( gct::html_color( c ) ),
       d, true, 0u, 1u, 0u
     );
@@ -329,10 +329,10 @@ int main( int argc, const char *argv[] ) {
         window.add_rectangle(
           vk::Rect2D{
             vk::Offset2D{
-              ( width - w )/2,
-              ( height - h )/2
+              static_cast< std::int32_t >( width - w )/2,
+              static_cast< std::int32_t >( height - h )/2
             },
-            vk::Extent2D{ w, h }
+            vk::Extent2D{ static_cast< std::uint32_t >( w ), static_cast< std::uint32_t >( h ) }
           },
           gct::srgb_oetf( gct::html_color( 0x000000 ) ),
           1u, true, 0u, 1u, 0u
